Use native ruby split in erb
The template in this module is manipulating data originally passed as a
string. The string data needs to be converted to an array. To achieve
this, the puppet parser function `split` was used. This did work but
will no longer be functional as the manner of calling these functions
has changed in Puppet 4. The Puppet parser split does not much more than
just expose the native ruby split to the Puppet DSL. In the ERB we have
full access to Ruby and do not need to invoke this internal parser
function.

This commit changes the method of splitting strings to use the native
ruby function. This change was tested to show that the original effect
of the ERB was still being achieved. This template is also now Puppet 4
ready, though once the module drops Puppet 3, EPP should be preferred
over ERB.