Fix service notifies for config changes
Currently a config file change will notify the proftpd service.
If service_manage is set to false a config change tries to notify the
service but the resource doesn't exist to notify.

Instead we should notify the service class so it will get passed down to
the resources in the class.
I also adjusted the notify to be more specific so it only notifies when
the actual config files change instead of every file in the class.